+++ doc/page-examples.h
@@ -1,0 +1,28 @@
+/**
+
+\page Examples
+
+\section BasicExample Basic example
+
+This example shows you how to use ivi-logging in a typical, using the standard configuration.
+
+logging-basic.h :
+\include "logging-basic.h"
+
+exampleLoggingMain.cpp :
+\include "exampleLoggingMain.cpp"
+
+
+
+\section FileExample Logging to a file
+
+This example shows you how to use ivi-logging in such a way that the logs are written to a file, in addition to the default channels.
+
+logging-file.h :
+\include "logging-file.h"
+
+exampleLoggingFile.cpp :
+\include "exampleLoggingFile.cpp"
+
+
+*/