+++ doc/page-faq.h
@@ -1,0 +1,13 @@
+/**
+
+\page FAQ
+
+\section FAQ FAQ
+
+\subsection File Logging to a file
+
+Q: How to log to a file.
+
+A: Check the "file" example in the examples folder.
+
+*/