+++ examples/basic/exampleLogging.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "logging.h"
+#include "logging-basic.h"
 
 // Import the application's contexts so that they can be used in any file which includes this header
 LOG_IMPORT_CONTEXT(mainContext);